requirements.txt
# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.18.1
google-cloud-logging==3.8.0

# API and HTTP
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0

# Cloud Functions
functions-framework==3.5.0